added a filter bar, cleaned up some js
