+++ app/assets/stylesheets/components/_index.scss
@@ -8,3 +8,4 @@
 @import "card-product";
 @import "map";
 @import "card";
+@import "filterbtn";
