+++ app/javascript/components/dropdown.js
@@ -1,0 +1,8 @@
+
+const dropdown = () => {
+
+
+}
+
+
+export { dropdown }
