+++ app/views/bookings/index.html.erb
@@ -2,7 +2,7 @@
 <div class="container mt-5">
   <br>
   <h1>My Bookings</h1>
-  <%= @bookings.each do |booking| %>
+  <% @bookings.each do |booking| %>
     <div class="card-product m-3">
       <img src=<%= cl_image_path booking.item.photo %>>
       <div class="card-product-infos">
