+++ app/views/items/index.html.erb
@@ -1,27 +1,8 @@
 <div class="margin">
-  <nav class="navbar navbar-dark bg-dark navbar-filter">
-    <a class="navbar-brand" href="#">Never expand</a>
-​
-    <div class="custom-control custom-switch mr-4" data-toggle="collapse" data-target="#map" aria-expanded="false" aria-controls="map">
-      <input type="checkbox" class="custom-control-input" id="customSwitch1">
-      <label class="custom-control-label text-white" for="customSwitch1">Close map</label>
-    </div>
-  </nav>
+  <%= render 'shared/filterbar' %>
   <div class="d-flex">
-    <div class="cards">
-      <% @items.each do |item| %>
-        <div class="card">
-          <img src='<%= cl_image_path item.photo, height: 300, width: 400, crop: :fill %>' class="card-img-top" alt="<%= item.name %>">
-          <div class="card-body">
-            <h5 class="card-text font-weight-lighter">
-              <%= link_to item.name, item_path(item), class: "stretched-link" %>
-              </h5>
-            <h6 class="card-text font-weight-normal"><%= item.description %></h6>
-            <p class="card-text mb-0"><%= item.price %>$ / per Night</p>
-            <p class="card-text">Address: <%= item.address %></p>
-          </div>
-        </div>
-      <% end %>
+    <div class="cards" id="cards-container">
+      <%= render 'items_partial', items: @items %>
     </div>
     <div
       id="map"
