+++ config/routes.rb
@@ -5,6 +5,8 @@ Rails.application.routes.draw do
   resources :items do
     resources :bookings, only: [:new, :create]
   end
+  get '/ascitems', to: 'items#filterasc'
+  get '/dscitems', to: 'items#filterdsc'
   resources :bookings, only: [:index, :show, :edit, :update, :destroy]
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
